Fixed a broken unit test

+++ core/Core/Variables/tests/VariableTest.cpp
@@ -116,7 +116,7 @@ void VariableTestFixture::testSimpleExpression() {
 	CPPUNIT_ASSERT(e->getValue().getFloat() == (float)(six * five));
 	
 	e = shared_ptr<ExpressionVariable>(new ExpressionVariable(&c1, &c2, M_DIVIDE));
-	CPPUNIT_ASSERT(e->getValue().getFloat() == (float)(six / five));
+	CPPUNIT_ASSERT(e->getValue().getFloat() == (double)(six / five));
 	
 	e = shared_ptr<ExpressionVariable>(new ExpressionVariable(&c1, &c2, M_MOD));
 	CPPUNIT_ASSERT(e->getValue().getInteger() == (long)(six % five));
